Specify rate in contract
The contract deployer now specifies a rate they expect the keypers to be
paid. The rate is given in tokens per second for all keypers together.
There is a function to check if the keypers have been paid this amount,
given the current block timestamp.

In addition, the contracts have been simplified slightly: Keypers are
specified in the constructor and it is not Ownable anymore.